requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
Pillow>=10.0
python-docx>=1.1
